requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data
Pillow>=9.5.0
kagglehub>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# MLflow
mlflow>=2.8.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# UI
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.0.0
